Project files refactoring. Import of TransactionRequest does not work.
