+++ transaction-service-api/src/main/java/pl/adamboguszewski/transactionservice/service/TransactionRequest.java
@@ -1,0 +1,24 @@
+package pl.adamboguszewski.transactionservice.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class TransactionRequest {
+
+    private final String id;
+
+    private final String name;
+
+    public TransactionRequest(String id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
