+++ transaction-service/src/main/java/pl/adamboguszewski/transactionservice/application/Application.java
@@ -1,0 +1,12 @@
+package pl.adamboguszewski.transactionservice.application;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
