+++ transaction-service/src/main/java/pl/adamboguszewski/transactionservice/application/ApplicationController.java
@@ -1,9 +1,6 @@
-package com.example.transactionservice.application;
-// [TODO]: Your package should be unique for every project but your group id should be same in every your project
-// [TODO]: Package name should be <group.id>.<project.name>.<module.name>.<directory.name>. It helps with trace error stack, recognise your projects/modules etc
-// [TODO]: Click PPM on package and select refactor > Rename... (shift+F6)
+package pl.adamboguszewski.transactionservice.application;
 
-import com.example.transactionservice.service.TransactionRequest;
+import pl.adamboguszewski.transactionservice.service.TransactionRequest;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
