--- transaction-service/src/test/java/com/example/transactionservice/TransactionServiceApplicationTests.java
@@ -1,17 +1,0 @@
-package com.example.transactionservice;
-// [TODO]: Your package should be unique for every project but your group id should be same in every your project
-// [TODO]: Package name should be <group.id>.<project.name>.<module.name>.<directory.name>. It helps with trace error stack, recognise your projects/modules etc
-// [TODO]: For test packages it should have .test ending
-// [TODO]: Click PPM on package and select refactor > Rename... (shift+F6)
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TransactionServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
