+++ transaction-service/src/test/java/pl/adamboguszewski/transactionservice/test/TransactionServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package pl.adamboguszewski.transactionservice.test;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TransactionServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
